Add files via upload
About Dataset
Data and descriptions are copy from http://kdd.ics.uci.edu/databases/kddcup99/kddcup99.html 

This is the data set used for The Third International Knowledge Discovery and Data Mining Tools Competition, which was held in conjunction with KDD-99 The Fifth International Conference on Knowledge Discovery and Data Mining. The competition task was to build a network intrusion detector, a predictive model capable of distinguishing between bad'' connections, called intrusions or attacks, andgood'' normal connections. This database contains a standard set of data to be audited, which includes a wide variety of intrusions simulated in a military network environment.

This code is a machine learning script that uses several libraries including pandas, numpy, sklearn, and statsmodels to perform several operations on a data file. The script first imports the necessary libraries and reads in a data file "data.csv" using pandas. The data is read in chunks to save memory and then concatenated into a single dataframe. The script then checks for null values in the dataframe, and if there are any, it will display the number of null values for each column. Next, the script separates the dataframe into two arrays, one for the independent variables (X) and one for the dependent variable (y). Next, the script uses LabelEncoder to encode the categorical values in the independent variable array. Then the script splits the data into training and testing sets using sklearn's train_test_split method. After that, the script applies standardization to the training and testing sets using sklearn's StandardScaler. After this, the script uses RandomForestRegressor, LogisticRegression, DecisionTreeRegressor and SVR to train on the training sets and make predictions on the testing sets. Finally, the script uses sklearn's explained_variance_score to evaluate the performance of the models.